+++ math.cpp
@@ -1,16 +1,16 @@
 #define PI 3.1415926535897932384626
 #define mod 1000000007
 //Mod function
-long long int md(long long int x)
+int md(int x)
 {
 	while(x<0)
 	x+=mod;
 	return x%mod;
 }
 //Multiply large Number num>=1e9 use Modulo
-long long int multi(long long int a, long long int b)
+int multi(int a, int b)
 {
-	long long int pro=0;
+	int pro=0;
 	a=md(a);
 	while(b)
 	{
@@ -22,7 +22,7 @@ long long int multi(long long int a, long long int b)
 	return pro;
 }
 //Power Function A raised to B use modulo
-long long int pwr(long long int a,long long int n)
+int pwr(int a,int n)
 {
 	if(n==0)
 	return 1;
@@ -30,7 +30,7 @@ long long int pwr(long long int a,long long int n)
 	return a;
 	else if(n&1)
 	{
-		long long int x=pwr(a,n/2);
+		int x=pwr(a,n/2);
 		if(x<mod)
 		return md(md(md(x)*md(x))*a);
 		else
@@ -38,7 +38,7 @@ long long int pwr(long long int a,long long int n)
 	}
 	else
 	{
-		long long int x=pwr(a,n/2);
+		int x=pwr(a,n/2);
 		if(x<mod)
 		return md(md(md(x)*md(x))*1);
 		else
@@ -46,31 +46,31 @@ long long int pwr(long long int a,long long int n)
 	}
 }
 //Modulo Inverse use modulo
-long long int ge(long long int a, long long int b, long long int *x, long long int *y)
+int ge(int a, int b, int *x, int *y)
 {
 	if (a == 0)
 	{
 		*x = 0, *y = 1;
 		return b;
 	}
-	long long int x1, y1;
-	long long int gcd = ge(b%a, a, &x1, &y1);
+	int x1, y1;
+	int gcd = ge(b%a, a, &x1, &y1);
 	*x = y1 - (b/a) * x1;
 	*y = x1;
 	return gcd;
 }
-long long int  modinv(long long int a, long long int m=mod)
+int  modinv(int a, int m=mod)
 {
-	long long int x, y;
-	long long int g = ge(a, m, &x, &y);
-	long long int p = (x%m + m) % m;
+	int x, y;
+	int g = ge(a, m, &x, &y);
+	int p = (x%m + m) % m;
 	return p;
 }
 //Prime Factor Generator
-vector <long long int> prmfct[MAX];
+vector <int> prmfct[MAX];
 void prmfactgenerator()
 {
-	long long int i,j;
+	int i,j;
 	for(i=2;i<MAX;i++)
 	{
 		if(!prmfct[i].size())
@@ -89,7 +89,7 @@ bool prime;
 void sieve()
 {
 	memset(prime,true,sizeof(prime));
-	long long int i,j;
+	int i,j;
 	for(i=2;i*i<=MAX;i++)
 	{
 		if(prime[i])
@@ -98,9 +98,9 @@ void sieve()
 	}
 }
 //Matrix Exponentation mat=2*2 use modulo
-void multex(long long int a[2][2],long long int c[2][2])
+void multex(int a[2][2],int c[2][2])
 {
-	long long int i,j,k,b[2][2]={0};
+	int i,j,k,b[2][2]={0};
 	for(i=0;i<2;i++)
 	for(j=0;j<2;j++)
 	for(k=0;k<2;k++)
@@ -112,9 +112,9 @@ void multex(long long int a[2][2],long long int c[2][2])
 	for(j=0;j<2;j++)
 	a[i][j]=b[i][j];
 }
-void matexp(long long int a[2][2],long long int ans[2][2],long long int s)
+void matexp(int a[2][2],int ans[2][2],int s)
 {
-	long long int i,j;
+	int i,j;
 	if(s==1)
 	{
 		ans[0][0]=a[0][0];
@@ -135,7 +135,7 @@ void matexp(long long int a[2][2],long long int ans[2][2],long long int s)
 	}
 }
 //nCr Matrix use modulo
-long long int ncr[MAX][MAX];
+int ncr[MAX][MAX];
 void ncomr()
 {
 	int i,j;
